% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_calc.R
\name{cor_calc}
\alias{cor_calc}
\title{Calculate correlation matrix with different methods and missing data handling}
\usage{
cor_calc(
  data,
  ordered = "adapted",
  missing_handling = "two-step-em",
  nimp = 20,
  imp_method = "pmm",
  max_categories = 7
)
}
\arguments{
\item{data}{A data frame or matrix containing the variables for which the correlation matrix is to be calculated.}

\item{ordered}{Specifies whether variables should be treated as ordered categorical when determining correlations.
Options are \code{TRUE}, \code{FALSE}, or \code{"adapted"}. The argument can be provided as a single value
(applied to all variables) or as a vector of length equal to the number of variables (using only \code{TRUE} and \code{FALSE}),
allowing mixed specifications. With the default \code{"adapted"}, the treatment of each variable is determined according to
guidelines from preliminary simulations (considering the number of cases, number of variables,
and number of categories).}

\item{missing_handling}{Method for estimating the correlation matrix in the presence of missing data.
\code{"tow-step-em"} uses a classic EM algorithm to estimate the correlation matrix from the data.
\code{"stacked-mi"} uses multiple imputation to estimate the correlation matrix from the data.
\code{"pairwise"} uses pairwise deletion to estimate the correlation matrix from the data.
\code{"listwise"} uses listwise deletion to estimate the correlation matrix from the data.}

\item{nimp}{Number of multiple imputations to perform when using multiple imputation for missing data (default: 20).}

\item{imp_method}{Method for multiple imputation when using \code{"stacked-mi"} for missing data handling. Default is \code{"pmm"} (predictive mean matching).}

\item{max_categories}{Maximum number of categories for a variable to be considered ordered if \code{cor_method = "adapted"} (default: 7).}
}
\value{
A list with the following elements:
\describe{
\item{mat}{Estimated correlation matrix.}
\item{missing_handling}{Method used for handling missing data (if applicable).}
\item{nimp}{Number of imputations used (if applicable).}
\item{cor_method}{Matrix indicating the correlation method used for each pair of variables.}
}
}
\description{
Calculate correlation matrix with different methods and missing data handling
}
\details{
While polychoric correlations are generally more appropriate for ordered categorical data, they rely on the assumption of multivariate normality
and may encounter estimation problems if the number of available observations is small relative to the number of estimated parameters. Preliminary
simulations suggest that in such cases Pearson correlations may introduce less bias, an effect that becomes even more pronounced when data are missing.

We therefore allow users either to specify the correlation type directly or to use an adaptive option that selects the method based on the simulation
results. In general, variables with more than seven categories can be treated as continuous, whereas for variables with fewer categories the procedure
evaluates whether the amount of available information is too limited to justify polychoric estimation, in which
case Pearson correlations are used instead. The method is applied pairwise: polychoric correlations are computed if both variables are treated as
ordinal, polyserial correlations if one variable is ordinal and the other continuous, and Pearson correlations if both are continuous. The adaptive
procedure is still under development and may be refined in future versions.
}
\examples{
# Estimate correlation matrix from full data set
result <- cor_calc(data = mantar_dummy_full, ordered = FALSE)

# View estimated correlation matrix and methods used
result$mat
result$cor_method

# Estimate correlation matrix for data set with missings
result_mis <- cor_calc(data = mantar_dummy_mis,
                      ordered = "adapted",
                      missing_handling = "two-step-em")

# View estimated correlation matrix and methods used
result_mis$mat
result_mis$cor_method
}
