% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_net.R
\name{neighborhood_net}
\alias{neighborhood_net}
\title{Estimate Network using Neighborhood Selection based on Information Criteria}
\usage{
neighborhood_net(
  data = NULL,
  ns = NULL,
  mat = NULL,
  n_calc = c("average", "individual", "max", "total"),
  missing_handling = c("two-step-em", "stacked-mi"),
  k = "log(n)",
  nimp = 20,
  pcor_merge_rule = c("and", "or")
)
}
\arguments{
\item{data}{Raw data containing only the variables to be included in the network. May include missing values.}

\item{ns}{Numeric vector specifying the sample size for each variable in the data.
If not provided, it will be computed based on the data.
Must be provided if a correlation matrix (\code{mat}) is supplied instead of raw data.}

\item{mat}{Optional covariance or correlation matrix for the variables to be included in the network.
Used only if \code{data} is \code{NULL}.}

\item{n_calc}{Method for calculating the sample size for node-wise regression models.}

\item{missing_handling}{Method for estimating the correlation matrix in the presence of missing data.}

\item{k}{Penalty per parameter (number of predictor + 1) to be used in node-wise regressions; the default log(n) (number of observations for the dependent variable) is the classical BIC. Alternitavely, classical AIC would be \code{k = 2}.}

\item{nimp}{Number of multiple imputations to perform when using multiple imputation for missing data (default: 20).}

\item{pcor_merge_rule}{Rule for merging regression weights into partial correlations.
\code{"and"} estimates a partial correlation only if regression weights in both directions (e.g., from node 1 to 2 and from 2 to 1) are non-zero in the final models.
\code{"or"} uses the available regression weight from one direction as partial correlation if the other is not included in the final model.}
}
\value{
A list with the following elements:
\describe{
\item{pcor}{Partial correlation matrix estimated from the node-wise regressions.}
\item{betas}{Matrix of regression coefficients from the final regression models.}
}
}
\description{
Estimate Network using Neighborhood Selection based on Information Criteria
}
\examples{
# LATER
}
