% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_regression.R
\name{regression_opt}
\alias{regression_opt}
\title{Title}
\usage{
regression_opt(
  data = NULL,
  n = NULL,
  mat = NULL,
  dep_ind,
  n_calc = c("average", "individual", "max", "total"),
  missing_handling = c("two-step-em", "stacked-mi"),
  k = "log(n)",
  nimp = 20
)
}
\arguments{
\item{data}{Raw data containing only the variables to be tested within the multiple regression as dependent or independent variable. May include missing values.}

\item{n}{Numeric value specifying the sample size used in calculating information criteria for model search.
If not provided, it will be computed based on the data.
If a correlation matrix (\code{mat}) is supplied instead of raw data, \code{n} must be provided.}

\item{mat}{Optional covariance or correlation matrix for the variables to be used within the multiple regression.
#' Used only if \code{data} is \code{NULL}.}

\item{dep_ind}{Index of the dependent variable in the regression model.}

\item{n_calc}{Method for calculating the sample size used in calculating information criteria for model search.}

\item{missing_handling}{Method for estimating the correlation matrix in the presence of missing data.}

\item{k}{Penalty per parameter (number of predictors + 1) to be used in node-wise regressions; the default log(n) (number of observations observation) is the classical BIC. Alternatively, classical AIC would be \code{k = 2}.}

\item{nimp}{Number of multiple imputations to perform when using multiple imputation for missing data (default: 20).}
}
\value{
A list with the following elements:
\describe{
\item{regression}{Named vector of regression coefficients for the dependent variable.}
\item{R2}{R-squared value of the regression model.}
}
}
\description{
Title
}
\examples{
# LATER
}
